Add imageAlign, imageArray and showTitle to top level nav json.  This allows user to use these at that level, allowing top nav level to be of type imageAndText or text.
Also configured text.html to always show title unless showTitle is false, and textAndImage.html to show title if showTitle is true.
Previously there was no ability to control on which type of pages the title showed.
The image.html page has not been changed, and will not show title regardless of the showTitle setting.    Will add that in next commit.